dtype issue fixed

+++ cs336_systems/benchmarking_scripts/benchmarking_script.py
@@ -46,8 +46,8 @@ dtype = torch.bfloat16 if args.bfloat16 else torch.float32
 batch_size = 4
 if torch.cuda.is_available():
     device = "cuda"
-random_input = torch.randint(low = 0, high = args.vocab_size, size = (batch_size, args.context_length), device=device, dtype=dtype)
-random_target = torch.randint(low = 0, high = args.vocab_size, size = (batch_size, args.context_length), device=device, dtype=dtype)
+random_input = torch.randint(low = 0, high = args.vocab_size, size = (batch_size, args.context_length), device=device)
+random_target = torch.randint(low = 0, high = args.vocab_size, size = (batch_size, args.context_length), device=device)
 model.to(device)
